move lang and theme resolvers into their own file
